refactor(ioxide): put the TLS handshake with its caller, drop the Tls folder
IoxideTls was a folder holding one four-line method with one caller, under the
prefix every other type has now shed. Its two halves belonged in different
places, which is why neither fit where it was:

  AcceptWithAlpnAsync is per-connection work on the reactor thread, and
  establishing a connection's transport is exactly what ConnectionDriver is for.
  It is now a private AcceptTlsAsync there, beside the plaintext branch it is the
  alternative to.

  TlsRegistry joins Server.Tcp.Tls.cs, where ResolveTls produces what fills it.

Not Server.Tcp.Tls for the handshake: that file is a partial of Server holding
startup configuration - instance members that run once per reactor in OnStart.
Terminating a connection has no server, and putting it there would have the
connection driver calling into Server for a transport primitive.
